@@ -125,22 +125,31 @@ func (x *XFile) Debugf(format string, v ...any) {
 	}
 }
 
-// GetFileList returns all the files in a path.
-// This is non-recursive and only returns files _in_ the base path provided.
+// GetFileList returns all the files in a path or paths.
+// This is non-recursive and only returns files _in_ the base paths provided.
 // This is a helper method and only exposed for convenience. You do not have to call this.
-func (x *Xtractr) GetFileList(path string) ([]string, error) {
-	if stat, err := os.Stat(path); err == nil && !stat.IsDir() {
-		return []string{path}, nil
-	}
+func (x *Xtractr) GetFileList(paths ...string) ([]string, error) {
+	files := []string{}
 
-	fileList, err := os.ReadDir(path)
-	if err != nil {
-		return nil, fmt.Errorf("reading path %s: %w", path, err)
-	}
+	for _, path := range paths {
+		stat, err := os.Stat(path)
+		if err != nil {
+			return nil, fmt.Errorf("stat: %w", err)
+		}
 
-	files := make([]string, len(fileList))
-	for idx, file := range fileList {
-		files[idx] = filepath.Join(path, file.Name())
+		if !stat.IsDir() {
+			files = append(files, path)
+			continue
+		}
+
+		fileList, err := os.ReadDir(path)
+		if err != nil {
+			return nil, fmt.Errorf("reading path %s: %w", path, err)
+		}
+
+		for _, file := range fileList {
+			files = append(files, filepath.Join(path, file.Name()))
+		}
 	}
 
 	return files, nil
@@ -149,7 +158,7 @@ func (x *Xtractr) GetFileList(path string) ([]string, error) {
 // Difference returns all the strings that are in slice2 but not in slice1.
 // Used to find new files in a file list from a path. ie. those we extracted.
 // This is a helper method and only exposed for convenience. You do not have to call this.
-func Difference(slice1 []string, slice2 []string) []string {
+func Difference(slice1, slice2 []string) []string {
 	diff := []string{}
 
 	for _, s2p := range slice2 {
@@ -263,7 +272,7 @@ func getCompressedFiles(path string, filter *Filter, fileList []os.FileInfo, dep
 				files[k] = v
 			}
 		case strings.HasSuffix(lowerName, ".rar"):
-			hasParts := regexp.MustCompile(`.*\.part[0-9]+\.rar$`)
+			hasParts := regexp.MustCompile(`.*\.part\d+\.rar$`)
 			partOne := regexp.MustCompile(`.*\.part0*1\.rar$`)
 			// Some archives are named poorly. Only return part01 or part001, not all.
 			if !hasParts.MatchString(lowerName) || partOne.MatchString(lowerName) {
@@ -282,13 +291,13 @@ func getCompressedFiles(path string, filter *Filter, fileList []os.FileInfo, dep
 
 // Extract calls the correct procedure for the type of file being extracted.
 // Returns size of extracted data, list of extracted files, and/or error.
-func (x *XFile) Extract() (int64, []string, []string, error) {
+func (x *XFile) Extract() (size int64, filesList, archiveList []string, err error) {
 	return ExtractFile(x)
 }
 
 // ExtractFile calls the correct procedure for the type of file being extracted.
 // Returns size of extracted data, list of extracted files, list of archives processed, and/or error.
-func ExtractFile(xFile *XFile) (int64, []string, []string, error) {
+func ExtractFile(xFile *XFile) (size int64, filesList, archiveList []string, err error) {
 	sName := strings.ToLower(xFile.FilePath)
 
 	for _, ext := range extension2function {
@@ -303,7 +312,7 @@ func ExtractFile(xFile *XFile) (int64, []string, []string, error) {
 // MoveFiles relocates files then removes the folder they were in.
 // Returns the new file paths.
 // This is a helper method and only exposed for convenience. You do not have to call this.
-func (x *Xtractr) MoveFiles(fromPath string, toPath string, overwrite bool) ([]string, error) { //nolint:cyclop
+func (x *Xtractr) MoveFiles(fromPath, toPath string, overwrite bool) ([]string, error) { //nolint:cyclop
 	var (
 		newFiles = []string{}
 		keepErr  error
@@ -403,22 +412,21 @@ func (x *Xtractr) Rename(oldpath, newpath string) error {
 	if err != nil {
 		return fmt.Errorf("os.Open(): %w", err)
 	}
+	defer oldFile.Close()
 
 	newFile, err := os.OpenFile(newpath, os.O_TRUNC|os.O_CREATE|os.O_WRONLY, x.config.FileMode)
 	if err != nil {
-		oldFile.Close()
 		return fmt.Errorf("os.OpenFile(): %w", err)
 	}
 	defer newFile.Close()
 
 	_, err = io.Copy(newFile, oldFile)
-	oldFile.Close()
-
 	if err != nil {
 		return fmt.Errorf("io.Copy(): %w", err)
 	}
 
 	// The copy was successful, so now delete the original file
+	_ = oldFile.Close() // Needs to be closed before delete.
 	_ = os.Remove(oldpath)
 
 	return nil
