@@ -11,7 +11,7 @@ import (
 /* How to extract a ZIP file. */
 
 // ExtractZIP extracts a zip file.. to a destination. Simple enough.
-func ExtractZIP(xFile *XFile) (int64, []string, error) {
+func ExtractZIP(xFile *XFile) (size int64, filesList []string, err error) {
 	zipReader, err := zip.OpenReader(xFile.FilePath)
 	if err != nil {
 		return 0, nil, fmt.Errorf("zip.OpenReader: %w", err)
@@ -19,15 +19,16 @@ func ExtractZIP(xFile *XFile) (int64, []string, error) {
 	defer zipReader.Close()
 
 	files := []string{}
-	size := int64(0)
+	size = int64(0)
 
-	for _, zipFile := range zipReader.Reader.File {
+	for _, zipFile := range zipReader.File {
 		fSize, wfile, err := xFile.unzip(zipFile)
 		if err != nil {
 			return size, files, fmt.Errorf("%s: %w", xFile.FilePath, err)
 		}
 
-		files = append(files, filepath.Join(xFile.OutputDir, zipFile.Name)) //nolint: gosec
+		//nolint:gosec // this is safe because we clean the paths.
+		files = append(files, filepath.Join(xFile.OutputDir, zipFile.Name))
 		size += fSize
 		xFile.Debugf("Wrote archived file: %s (%d bytes), total: %d files and %d bytes", wfile, fSize, len(files), size)
 	}
